% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{glmnet_analysis}
\alias{glmnet_analysis}
\title{Machine Learning Based on regressive analysis for evaluate model}
\usage{
glmnet_analysis(data, cv_k)
}
\arguments{
\item{data}{A data frame containing the dependent variable, which should be a numeric on the last column, with the column name of "y".}

\item{cv_k}{The K-fold cross validation method divides the proportion of training and validation sets, such as 5 representing 4 training sets and 1 validation set.}
}
\value{
A plot of actual and predicted values of dependent variable after each cross validation; A data frame of mse, mae, rmse, R2 (including training and validation sets) after each cross validation.
}
\description{
Machine Learning Based on regressive analysis for evaluate model
}
\examples{
data <- MASS::Boston; colnames(data)[ncol(data)] <- "y"; glmnet_analysis(data=data,cv_k=5)
}
