% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_union.R
\name{ML_analysis}
\alias{ML_analysis}
\title{Machine Learning Based on xgboost, random forest, and mlp neural network}
\usage{
ML_analysis(data, prop, var_num)
}
\arguments{
\item{data}{A data frame containing the dependent variable, which should be a factor, with the column name of "y".}

\item{prop}{The proportion of dividing data into training and validation sets.}

\item{var_num}{Number of independent variables in the data.}
}
\value{
For regression models, return the performance evaluation list and true/predicted value match plot for each model; For classification models, return the performance evaluation list and ROC curve for each model.
}
\description{
Machine Learning Based on xgboost, random forest, and mlp neural network
}
\examples{
colnames(iris)[1] <- "y"; ML_analysis(data=iris,prop=0.7,var_num=4)
}
