% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{na_process}
\alias{na_process}
\title{Handling missing values}
\usage{
na_process(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame which had no missing value
}
\description{
Handling missing values
}
\examples{
na_process(data=iris)
}
