% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{num_plot}
\alias{num_plot}
\title{Box plot of numeric variables}
\usage{
num_plot(
  num,
  group,
  width = 0.5,
  xlab = "variable",
  ylab = "",
  grouplab = "Group",
  color = c("#2874C5", "#f87669", "#e6b707", "#868686", "#66C2A5", "#FC8D62", "#8DA0CB",
    "#E78AC3", "#A6D854", "#FFD92F", "#E5C494", "#B3B3B3")
)
}
\arguments{
\item{num}{a data frame consist of numeric variables}

\item{group}{A factor with duplicated character or factor, which must correspond to the obs in num}

\item{width}{width of boxplot and error bar}

\item{xlab}{title of the x axis}

\item{ylab}{title of the y axis}

\item{grouplab}{title of the group legend}

\item{color}{color vector}
}
\value{
a boxplot according to \code{num} and grouped by \code{group}.
}
\description{
Box plot of numeric variables
}
\examples{
num_plot(num=iris[,1:4],group=iris[,5])
}
