% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{outlier_process}
\alias{outlier_process}
\title{Outlier processing for data without missing values}
\usage{
outlier_process(data)
}
\arguments{
\item{data}{a data frame with no missing value, which consist of character or numerical variables}
}
\value{
a data frame without outlier, especially for numeric column
}
\description{
Outlier processing for data without missing values
}
\examples{
iris$Species <- as.character(iris$Species); outlier_process(data=iris)
}
