% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{rf_analysis}
\alias{rf_analysis}
\title{Machine Learning Based on Random Forest for Selecting Variables}
\usage{
rf_analysis(data, cv_k)
}
\arguments{
\item{data}{A data frame containing the dependent variable, which should be a factor, with the column name of "y".}

\item{cv_k}{The K-fold cross validation method divides the proportion of training and validation sets, such as 5 representing 4 training sets and 1 validation set.}
}
\value{
A plot of contribution of random forest variables to the model; A data frame of accuracy, F1 score (including training and validation sets) after each cross validation.
}
\description{
Machine Learning Based on Random Forest for Selecting Variables
}
\examples{
colnames(iris)[5] <- "y"; rf_analysis(data=iris,cv_k=4)

}
