% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{vec_plot}
\alias{vec_plot}
\title{Bar plot of categorical variables}
\usage{
vec_plot(
  vec,
  group,
  color = c("#2874C5", "#f87669", "#e6b707", "#868686", "#66C2A5", "#FC8D62", "#8DA0CB",
    "#E78AC3", "#A6D854", "#FFD92F", "#E5C494", "#B3B3B3")
)
}
\arguments{
\item{vec}{a data frame consist of categorical variables}

\item{group}{A factor with duplicated character or factor, which must correspond to the obs in vec}

\item{color}{color vector}
}
\value{
a barplot according to \code{vec} and grouped by \code{group}.
}
\description{
Bar plot of categorical variables
}
\examples{
vec_plot(vec=mtcars[,1:10],group=mtcars[,10])
}
